requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-tastypie==0.14.2
idna==2.8
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
